pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "auth>=0.5.3",
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.39",
    "werkzeug>=3.1.3",
    "pillow>=11.1.0",
    "face-recognition>=1.3.0",
    "wtforms>=3.2.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
]
